Added AI Targeting
AI character has a sphere trigger that adds any characters that moves into the sphere to a list. If the AI has a line of sight using a raycast the enemy will set the object in sight as its target. The AI will set its destination as the targets position.