Added Win Objective
Created a new trigger target that will end the game when the correct objective is given to it.
The win trigger also has rotating cube around it to show where it is.
Also removed Item Trigger script as it was unnecessary and replaced it with setting the required gameobject on the item trigger.