AI Character Now Checks if Firing Line is Clear
When the AI is in range of its target the AI checks to see if any obstacles are in the way, if there are it continues to path find to its target other wise the AI starts attacking the target.
The AI will also stop attacking and continue path finding to the target if the distance is too large.